working racing phase
notes: the cards all seem to be "or" cards, despite checking for "and" cards. i'm pretty sure this is due to how we make cards with the builder, we should add a "setOrStatus" method or something to the builder to allow cards made in the deck generator to be and cards.